feat: 전체 이슈 갯수 및 백로그 이슈 갯수 구현

+++ backend/core/core-api/src/main/java/com/ssafy/taskit/api/controller/EpicResponse.java
@@ -5,16 +5,14 @@ import java.util.List;
 import java.util.Map;
 
 public record EpicResponse(
-    Long id, String key, String name, Integer cntTotalIssues, Integer cntRemainIssues) {
+    Long id, String key, String name, Long cntTotalIssues, Long cntRemainIssues) {
 
   public static List<EpicResponse> from(
-      List<Epic> epics,
-      Map<Long, Integer> totalIssueCountMap,
-      Map<Long, Integer> backlogIssueCountMap) {
+      List<Epic> epics, Map<Long, Long> totalIssueCountMap, Map<Long, Long> backlogIssueCountMap) {
     return epics.stream()
         .map(epic -> {
-          Integer cntTotalIssues = totalIssueCountMap.getOrDefault(epic.id(), 0);
-          Integer cntRemainIssues = backlogIssueCountMap.getOrDefault(epic.id(), 0);
+          Long cntTotalIssues = totalIssueCountMap.getOrDefault(epic.id(), 0L);
+          Long cntRemainIssues = backlogIssueCountMap.getOrDefault(epic.id(), 0L);
 
           return new EpicResponse(
               epic.id(), epic.key(), epic.name(), cntTotalIssues, cntRemainIssues);
